+++ tests/__init__.py
@@ -1,6 +1,6 @@
 import unittest
 
-from elex.api import Election, DelegateReport, utils
+from elex.api import Election, DelegateReport, USGovernorTrendReport, USHouseTrendReport, USSenateTrendReport, utils
 from time import sleep
 
 API_MESSAGE = "We require that you export AP_API_KEY in your environment in order to test AP connectivity."
@@ -16,6 +16,17 @@ class DelegateReportTestCase(unittest.TestCase):
             delsum_datafile=self.delsum_datafile
         )
         self.delegate_reports = d.candidate_objects
+
+
+class TrendReportTestCase(unittest.TestCase):
+    governor_file = 'tests/data/20160818_gov_trends.json'
+    house_file = 'tests/data/20160818_house_trends.json'
+    senate_file = 'tests/data/20160818_senate_trends.json'
+
+    def setUp(self, **kwargs):
+        self.governor_trends = USGovernorTrendReport(self.governor_file)
+        self.house_trends = USHouseTrendReport(self.house_file)
+        self.senate_trends = USSenateTrendReport(self.senate_file)
 
 
 class ElectionResultsParseIdsTestCase(unittest.TestCase):